refactor: move projection and intersection to curves submodule
