+++ src/pynurbs/__init__.py
@@ -1,6 +1,5 @@
 from .basis_functions import BasisFunctions
 from .knotspace import GeneratorKnotVector, KnotVector
-from .operations.advanced import Intersection, Projection
 from .operations.calculus import Derivate, Integrate
 
 __version__ = "1.1.0"
